requirements.txt
feedparser==6.0.11
anthropic==0.49.0
gspread==6.0.2
google-auth==2.29.0
beautifulsoup4==4.12.3
requests==2.31.0
praw==7.7.1
